backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
pydantic-settings
pydantic[email]
sqlalchemy
psycopg2-binary
celery[redis]
redis
playwright
beautifulsoup4
bcrypt
requests
alembic
stem
structlog
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-celery
authlib
python-dotenv
aiosmtplib
jinja2
email-validator
httpx
motor

#testes
pytest-asyncio